# [java] Don't use ClassNotFoundException to report unresolved classes