--- a/pmd-java/src/main/java/net/sourceforge/pmd/lang/java/typeresolution/PMDASMClassLoader.java
+++ b/pmd-java/src/main/java/net/sourceforge/pmd/lang/java/typeresolution/PMDASMClassLoader.java
@@ -30,7 +30,7 @@ import net.sourceforge.pmd.lang.java.typeresolution.visitors.PMDASMVisitor;
  * the negative cases only. The cache is shared between loadClass and getImportedClasses,
  * as they are using the same (parent) class loader, e.g. if the class foo.Bar cannot be loaded,
  * then the resource foo/Bar.class will not exist, too.
- * 
+ *
  * Note: since git show 46ad3a4700b7a233a177fa77d08110127a85604c the cache is using
  * a concurrent hash map to avoid synchronizing on the class loader instance.
  */
@@ -70,21 +70,27 @@ public final class PMDASMClassLoader extends ClassLoader {
 
     @Override
     public Class<?> loadClass(String name) throws ClassNotFoundException {
-        if (dontBother.containsKey(name)) {
+        Class<?> aClass = loadClassOrNull(name);
+        if (aClass == null) {
             throw new ClassNotFoundException(name);
+        }
+        return aClass;
+    }
+
+    /**
+     * Not throwing CNFEs to represent failure makes a huge performance
+     * difference. Typeres as a whole is 2x faster.
+     */
+    public Class<?> loadClassOrNull(String name) {
+        if (dontBother.containsKey(name)) {
+            return null;
         }
 
         try {
             return super.loadClass(name);
-        } catch (ClassNotFoundException e) {
+        } catch (ClassNotFoundException | LinkageError e) {
             dontBother.put(name, Boolean.TRUE);
-            throw e;
-        } catch (NoClassDefFoundError e) {
-            dontBother.put(name, Boolean.TRUE);
-            // rethrow as ClassNotFoundException, as the remaining part just
-            // deals with that
-            // see also: https://sourceforge.net/p/pmd/bugs/1319/
-            throw new ClassNotFoundException(name, e);
+            return null;
         }
     }
 
@@ -93,7 +99,7 @@ public final class PMDASMClassLoader extends ClassLoader {
      * doesn't know for sure it will fail). Notice, that the ability to resolve
      * a class does not imply that the class will actually be found and
      * resolved.
-     * 
+     *
      * @param name
      *            the name of the class
      * @return whether the class can be resolved