--- a/pmd-java/src/main/resources/category/java/security.xml
+++ b/pmd-java/src/main/resources/category/java/security.xml
@@ -53,7 +53,7 @@ public class Foo {
     void bad() {
         byte[] iv = new byte[] { 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, };
     }
-    
+
     void alsoBad() {
         byte[] iv = "secret iv in here".getBytes();
     }